release: fusedtok 1.5.2 (#48)
Audit-driven hardening, round three over the previously-excluded
sections: the batched radix rounds now apply the row's repetition
penalty (the unpenalized selection prefix mis-composed
decode_step_batched's window under penalty != 1 - the widening loop
self-healed every draw, which is why five hundred tests stayed
green); the kv_append_paged span guard is 64-bit; duplicate penalty
ids no longer diverge between CPU and GPU; rope grids, launcher
guards and the batched CPU references got the checks their siblings
have; docs finish the deferred style sweep and the 1.5.1
verification findings. 527 tests on both GPUs; no API or benchmark
changes.

+++ src/rope.cu
@@ -146,8 +146,10 @@ rope_neox_cpu(const std::vector<float>& q, const std::vector<float>* k,
 void rope_launch(const float* x, float* y, int seq, int dim, float theta,
                  int pos_offset, std::uintptr_t stream) {
     if (seq <= 0 || dim <= 0) return;
-    int pairs = seq * (dim / 2);
-    rope_kernel<float><<<(pairs + kBlock - 1) / kBlock, kBlock, 0, (cudaStream_t)stream>>>(
+    // 64-bit: seq * pairs can exceed INT_MAX for long caches
+    long long pairs = (long long)seq * (dim / 2);
+    rope_kernel<float><<<(unsigned)((pairs + kBlock - 1) / kBlock), kBlock,
+        0, (cudaStream_t)stream>>>(
         x, y, seq, dim, theta, pos_offset);
     check_launch("rope kernel launch");
 }
@@ -155,8 +157,10 @@ void rope_launch(const float* x, float* y, int seq, int dim, float theta,
 void rope_launch_bf16(const __nv_bfloat16* x, __nv_bfloat16* y, int seq,
                       int dim, float theta, int pos_offset, std::uintptr_t stream) {
     if (seq <= 0 || dim <= 0) return;
-    int pairs = seq * (dim / 2);
-    rope_kernel<__nv_bfloat16><<<(pairs + kBlock - 1) / kBlock, kBlock, 0, (cudaStream_t)stream>>>(
+    // 64-bit: seq * pairs can exceed INT_MAX for long caches
+    long long pairs = (long long)seq * (dim / 2);
+    rope_kernel<__nv_bfloat16><<<(unsigned)((pairs + kBlock - 1) / kBlock),
+        kBlock, 0, (cudaStream_t)stream>>>(
         x, y, seq, dim, theta, pos_offset);
     check_launch("rope bf16 kernel launch");
 }
@@ -164,9 +168,10 @@ void rope_launch_bf16(const __nv_bfloat16* x, __nv_bfloat16* y, int seq,
 void rope_neox_launch(const float* x, float* y, int seq, int dim, float theta,
                       int pos_offset, std::uintptr_t stream) {
     if (seq <= 0 || dim <= 0) return;
-    int threads_needed = seq * (dim / 2);
+    // 64-bit: seq * pairs can exceed INT_MAX for long caches
+    long long threads_needed = (long long)seq * (dim / 2);
     rope_neox_kernel<float>
-        <<<(threads_needed + kBlock - 1) / kBlock, kBlock, 0,
+        <<<(unsigned)((threads_needed + kBlock - 1) / kBlock), kBlock, 0,
            (cudaStream_t)stream>>>(
             x, y, seq, dim, theta, pos_offset);
     check_launch("rope_neox kernel launch");
@@ -175,9 +180,10 @@ void rope_neox_launch(const float* x, float* y, int seq, int dim, float theta,
 void rope_neox_launch_bf16(const __nv_bfloat16* x, __nv_bfloat16* y, int seq,
                            int dim, float theta, int pos_offset, std::uintptr_t stream) {
     if (seq <= 0 || dim <= 0) return;
-    int threads_needed = seq * (dim / 2);
+    // 64-bit: seq * pairs can exceed INT_MAX for long caches
+    long long threads_needed = (long long)seq * (dim / 2);
     rope_neox_kernel<__nv_bfloat16>
-        <<<(threads_needed + kBlock - 1) / kBlock, kBlock, 0,
+        <<<(unsigned)((threads_needed + kBlock - 1) / kBlock), kBlock, 0,
            (cudaStream_t)stream>>>(
             x, y, seq, dim, theta, pos_offset);
     check_launch("rope_neox bf16 kernel launch");
